smash_reader/requirements.txt
mss==4.0.2
numpy==1.16.2
opencv-python==4.1.0.25
Pillow==6.0.0
pytesseract==0.2.6
